% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Asset}
\alias{Asset}
\title{Create a new Asset.}
\usage{
Asset(
  rid,
  file,
  name = NULL,
  type = NULL,
  description = NULL,
  tags = list(),
  metadata = list()
)
}
\arguments{
\item{rid}{Resource id.}

\item{file}{File path.}

\item{name}{Name.}

\item{type}{Type.}

\item{description}{Description.}

\item{tags}{List of tags.}

\item{metadata}{List of metadata.}
}
\description{
Create a new Asset.
}
