% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{asset_from_json}
\alias{asset_from_json}
\title{Converts a list of properties to an Asset.
The list should mirror the structure of the JSON representation of an Asset.}
\usage{
asset_from_json(asset)
}
\arguments{
\item{asset}{List of properties.}
}
\value{
Asset.
}
\description{
Converts a list of properties to an Asset.
The list should mirror the structure of the JSON representation of an Asset.
}
