% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{children}
\alias{children}
\title{Gets the children of the Container.}
\usage{
children(db, container)
}
\arguments{
\item{db}{Syre database connection.}

\item{container}{Syre Container.}
}
\value{
List of the Container's children.
}
\description{
Gets the children of the Container.
}
\examples{
db <- database()
container <- db |> find_container(type = "child")
childs <- db |> children(container)
}
