% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{Container}
\alias{Container}
\title{Create a new Container.}
\usage{
Container(
  rid,
  name = NULL,
  type = NULL,
  description = NULL,
  tags = list(),
  metadata = list(),
  assets = list()
)
}
\arguments{
\item{rid}{Resource id.}

\item{name}{Name.}

\item{type}{Type.}

\item{description}{Description.}

\item{tags}{List of tags.}

\item{metadata}{List of metadata.}

\item{assets}{List of Assets.}
}
\description{
Create a new Container.
}
