% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{container_from_search_json}
\alias{container_from_search_json}
\title{Converts a list of properties from a container query into a Container.
The list should mirror the structure of the JSON representation of a Container.}
\usage{
container_from_search_json(container)
}
\arguments{
\item{container}{List of properties.}
}
\value{
Container.
}
\description{
Converts a list of properties from a container query into a Container.
The list should mirror the structure of the JSON representation of a Container.
}
