% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{create_database}
\alias{create_database}
\title{Return whether Thot is running in development mode.}
\usage{
create_database()
}
\description{
Return whether Thot is running in development mode.
}
