% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{database}
\alias{database}
\title{Create a new Syre database connection.}
\usage{
database(dev_root = NULL, chdir = TRUE)
}
\arguments{
\item{dev_root}{Interactive graph root.
Used to set the development root when using a script interactively.
The script will set its graph root to the \code{Container} at the given path.
This is ignored if the script is being run by a runner.}

\item{chdir}{Change the working directory to the analysis root.
Defaults to \code{True}.}
}
\value{
A Syre Database connection.
}
\description{
Create a new Syre database connection.
}
\examples{
db <- database(dev_root = "/path/to/my/syre/project/container")
}
