% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{database_available}
\alias{database_available}
\title{Checks if a Syre database is available.}
\usage{
database_available()
}
\value{
Whether a Syre database is available.
}
\description{
Checks if a Syre database is available.
}
