% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{database_dev}
\alias{database_dev}
\title{Initialize a database in a development environment.}
\usage{
database_dev(socket, dev_root, chdir)
}
\arguments{
\item{socket}{ZMQ socket to use.}

\item{dev_root}{Absolute system path to root container for database.}

\item{chdir}{Change directory to project analyses folder.}
}
\description{
Initialize a database in a development environment.
}
