% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{database_prod}
\alias{database_prod}
\title{Initialize a database in a production environment.}
\usage{
database_prod(socket, project, root, chdir)
}
\arguments{
\item{socket}{ZMQ socket to use.}

\item{project}{Project id.}

\item{root}{Root container graph path.}

\item{chdir}{Change directory to project analyses folder.}
}
\description{
Initialize a database in a production environment.
}
