% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{database_server_path}
\alias{database_server_path}
\title{Get the path to the correct database server.}
\usage{
database_server_path()
}
\value{
Path to the local database executable for the current system.
}
\description{
Get the path to the correct database server.
}
