% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{escape_str}
\alias{escape_str}
\title{Escape special string characters.
Characters escaped:
\itemize{
\item Backslash (\)
}}
\usage{
escape_str(path)
}
\description{
Escape special string characters.
Characters escaped:
\itemize{
\item Backslash (\)
}
}
