% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{find_assets}
\alias{find_assets}
\title{Find Assets matching the given filter criteria.}
\usage{
find_assets(db, name = NULL, type = NULL, tags = NULL, metadata = NULL)
}
\arguments{
\item{db}{Thot database connection.}

\item{name}{Name of the Asset to match.}

\item{type}{Type of the Asset to match.}

\item{tags}{List of tags the Asset has to match.}

\item{metadata}{Named list of metadata the Asset has to match.}
}
\value{
List of Assets matching the filter.
}
\description{
Find Assets matching the given filter criteria.
}
\examples{
db <- database()
assets <- find_assets(db, type = "my_asset")
}
