% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{find_container}
\alias{find_container}
\title{Finds a single Container matching the given filter criteria.
If multiple matching Containers are found, a random one is returned.}
\usage{
find_container(db, name = NULL, type = NULL, tags = NULL, metadata = NULL)
}
\arguments{
\item{db}{Thot database connection.}

\item{name}{Name of the Container to match.}

\item{type}{Type of the Container to match.}

\item{tags}{List of tags the Container has to match.}

\item{metadata}{Named list of metadata the Container has to match.}
}
\value{
Single Container matched by the filter or \code{NULL} if none exist.
}
\description{
Finds a single Container matching the given filter criteria.
If multiple matching Containers are found, a random one is returned.
}
\examples{
db <- database()
container <- find_container(db, name = "My Container")
}
