% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{find_containers}
\alias{find_containers}
\title{Find Containers matching the given filter criteria.}
\usage{
find_containers(db, name = NULL, type = NULL, tags = NULL, metadata = NULL)
}
\arguments{
\item{db}{Thot database connection.}

\item{name}{Name of the Container to match.}

\item{type}{Type of the Container to match.}

\item{tags}{List of tags the Container has to match.}

\item{metadata}{Named list of metadata the Container has to match.}
}
\value{
List of Containers matching the filter.
}
\description{
Find Containers matching the given filter criteria.
}
\examples{
db <- database()
containers <- find_containers(db, type = "my_container")
}
