% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{flag}
\alias{flag}
\title{Flags a resource (Container or Asset).}
\usage{
flag(db, resource, message, severity = c("warning", "info", "error"))
}
\arguments{
\item{db}{Syre database connection.}

\item{resource}{Resource to flag.}

\item{message}{Message to display.}

\item{severity}{Flag's everity.}
}
\description{
Flags a resource (Container or Asset).
}
\examples{
db <- database()
asset <- db |> find_asset()
db |> flag(asset, "Check me!")
}
