% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{get_container}
\alias{get_container}
\title{Get a container.}
\usage{
get_container(socket, project, container)
}
\arguments{
\item{socket}{ZMQ socket.}

\item{project}{Project id.}

\item{container}{Graph path to project. (i.e. Root container has path \code{/}.)}
}
\value{
List of Container properties.
}
\description{
Get a container.
}
