% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{join_path_windows}
\alias{join_path_windows}
\title{Normalize Windows paths.
If total path length is less than 256, returns the normalized path,
otherwise returns the device (UNC) path.}
\usage{
join_path_windows(p1, p2)
}
\arguments{
\item{p1}{Base path.}

\item{p2}{Extension path.}
}
\value{
Normalized path.
}
\description{
Normalize Windows paths.
If total path length is less than 256, returns the normalized path,
otherwise returns the device (UNC) path.
}
