% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{load_project}
\alias{load_project}
\title{Loads a Syre project.}
\usage{
load_project(socket, path)
}
\arguments{
\item{socket}{ZMQ socket.}

\item{path}{Path to the Project.}
}
\value{
List of Project properties.
}
\description{
Loads a Syre project.
}
