% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{parent,Database,Asset-method}
\alias{parent,Database,Asset-method}
\title{Gets the parent of the Asset.}
\usage{
\S4method{parent}{Database,Asset}(db, resource)
}
\arguments{
\item{db}{Syre database connection.}

\item{resource}{Syre Asset.}
}
\value{
Asset's parent Container.
}
\description{
Gets the parent of the Asset.
}
\examples{
db <- database()
asset <- db |> find_asset(type = "data")
container <- db |> parent(asset)
}
