% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{parent,Database,Container-method}
\alias{parent,Database,Container-method}
\title{Gets the parent of the Container within the database context.}
\usage{
\S4method{parent}{Database,Container}(db, resource)
}
\arguments{
\item{db}{Thot database connection.}

\item{resource}{Thot Container.}
}
\value{
Container's parent or \code{NULL} if the root of the database.
}
\description{
Gets the parent of the Container within the database context.
}
\examples{
db <- database()
container <- db |> find_container(type = "child")
parent <- db |> parent(container)
}
