% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/resources.R
\name{parent}
\alias{parent}
\title{Get the parent of a resource.}
\usage{
parent(db, resource)
}
\arguments{
\item{db}{Thot database connection.}

\item{resource}{Thot resource.}
}
\value{
Parent of the resource, or \code{NULL} if it does not exist in the current context.
}
\description{
Get the parent of a resource.
}
