% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{project_resource_root_path}
\alias{project_resource_root_path}
\title{Gets the Project path given a path.
Returns \code{NULL} if the path is not in a project.}
\usage{
project_resource_root_path(path)
}
\arguments{
\item{path}{Path to get the Project root of.}
}
\value{
Project path of the resource, or \code{NULL}.
}
\description{
Gets the Project path given a path.
Returns \code{NULL} if the path is not in a project.
}
