% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{send_cmd}
\alias{send_cmd}
\title{Send a command to the database.}
\usage{
send_cmd(socket, cmd, result = TRUE)
}
\arguments{
\item{socket}{ZMQ socket.}

\item{cmd}{Command to send.}

\item{result}{Handle response as a Result.}
}
\value{
Deserialzed result.
}
\description{
Send a command to the database.
}
