% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/common.R
\name{split_path}
\alias{split_path}
\title{Split a path into its parts.}
\usage{
split_path(path)
}
\arguments{
\item{path}{Path to split.}
}
\value{
List of path parts.
}
\description{
Split a path into its parts.
}
