% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StandardProperties}
\alias{StandardProperties}
\title{Create a new \code{StandardProperties} list.}
\usage{
StandardProperties()
}
\value{
Default standard properties as a named list.
}
\description{
Create a new \code{StandardProperties} list.
}
