% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/syre-package.R
\docType{package}
\name{syre-package}
\alias{syre}
\alias{syre-package}
\title{syre: Syre scentific data management and analysis bindings for R.}
\description{
R bindings to interact with the Syre scientific data management and analysis tool suite.
}
\author{
\strong{Maintainer}: Brian Carlsen \email{carlsen.bri@gmail.com}

}
\keyword{internal}
