% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{syre_analysis_id}
\alias{syre_analysis_id}
\title{Gets the \code{SYRE_ANALYSIS_ID} environment variable.}
\usage{
syre_analysis_id()
}
\value{
Analysis id or \code{NA}.
}
\description{
Gets the \code{SYRE_ANALYSIS_ID} environment variable.
}
