% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{syre_container_id}
\alias{syre_container_id}
\title{Gets the \code{SYRE_CONTAINER_ID} environment variable.}
\usage{
syre_container_id()
}
\value{
Active Container id or \code{NA}.
}
\description{
Gets the \code{SYRE_CONTAINER_ID} environment variable.
}
