% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{syre_container_path}
\alias{syre_container_path}
\title{Gets the \code{SYRE_CONTAINER_ID} environment variable.}
\usage{
syre_container_path()
}
\value{
Root container path or \code{NA}.
}
\description{
Gets the \code{SYRE_CONTAINER_ID} environment variable.
}
