% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{syre_project_id}
\alias{syre_project_id}
\title{Gets the \code{SYRE_PROJECT_ID} environment variable.}
\usage{
syre_project_id()
}
\value{
Project id or \code{NA}.
}
\description{
Gets the \code{SYRE_PROJECT_ID} environment variable.
}
