% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/thot-package.R
\docType{package}
\name{thot-package}
\alias{thot}
\alias{thot-package}
\title{thot: Thot data management and analysis bindings for R.}
\description{
R bindings to interact with the Thot data management and analysis tool suite.
}
\author{
\strong{Maintainer}: Brian Carlsen \email{carlsen.bri@gmail.com}

}
\keyword{internal}
