% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api.R
\name{database}
\alias{database}
\title{Create a new Syre database connection.}
\usage{
database(dev_root = NULL)
}
\arguments{
\item{dev_root}{Path to the root Container for use in development mode.}
}
\value{
A Syre Database connection.
}
\description{
Create a new Syre database connection.
}
\examples{
db <- database(dev_root = "/path/to/my/syre/project/container")
}
