% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{database}
\alias{database}
\title{Create a new Thot database connection.}
\usage{
database(dev_root = NULL)
}
\arguments{
\item{dev_root}{Path to the root Container for use in development mode.}
}
\value{
A Thot Database connection.
}
\description{
Create a new Thot database connection.
}
\examples{
db <- database(dev_root = "/path/to/my/thot/project/container")
}
