% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/commands.R
\name{database_available}
\alias{database_available}
\title{Checks if a Thot database is available.}
\usage{
database_available()
}
\arguments{
\item{socket}{ZMQ socket.}
}
\value{
Whether a Thot database is available.
}
\description{
Checks if a Thot database is available.
}
