% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extendr-wrappers.R
\name{find_containers}
\alias{find_containers}
\title{Return whether Thot is running in development mode.}
\usage{
find_containers()
}
\description{
Return whether Thot is running in development mode.
}
