% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{StandardProperties}
\alias{StandardProperties}
\title{Create a new \code{StandardProperties} list.}
\usage{
StandardProperties(name = NULL, type = NULL, tags = list(), metadata = list())
}
\arguments{
\item{name}{Name of the Asset to match.}

\item{type}{Type of the Asset to match.}

\item{tags}{List of tags the Asset has to match.}

\item{metadata}{Named list of metadata the Asset has to match.}
}
\value{
Standard properties as a named list.
}
\description{
Create a new \code{StandardProperties} list.
}
