% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{thot_container_id}
\alias{thot_container_id}
\title{Gets the \code{THOT_CONTAINER_ID} environment variable.}
\usage{
thot_container_id()
}
\value{
Active Container id or \code{NA}.
}
\description{
Gets the \code{THOT_CONTAINER_ID} environment variable.
}
